% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatoKrige.R
\name{climatoKrige}
\alias{climatoKrige}
\title{climatoKrige}
\usage{
climatoKrige(d, xmin, ymin, xmax, ymax, taille, periode)
}
\description{
Krigeage selon un parametre climato, une bounding box, une taille de cellule et une periode
}
\examples{
climatoKrige("NORTAV", 765366.9039,6365089.3268,862527.7841,6445434.0002,3000,1)
}
