% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperatures-data.R
\docType{data}
\name{Temperatures}
\alias{Temperatures}
\title{Indices annuels de temperature et nombre de jours de temperature issus du modele Aladin-Climat}
\format{An object of class \code{"data.frame"}}
\source{
\href{https://www.data.gouv.fr/s/resources/indices-annuels-de-temperature-et-nombre-de-jours-de-temperature-issus-du-modele-aladin-pour-la-periode-de-reference/20150930-200648/Tempe_REF_annuel.txt}{https://www.data.gouv.fr/s/resources/indices-annuels-de-temperature-et-nombre-de-jours-de-temperature-issus-du-modele-aladin-pour-la-periode-de-reference/20150930-200648/Tempe_REF_annuel.txt}
}
\usage{
Temperatures
}
\description{
Moyennes des differents indices annuels de temperatures et de nombre de jours de temperatures, sur les 8602 points terrestres de la France metropolitaine sur la periode de reference (1976-2005)
Moyennes calculees à partir de donnees corrigees du CNRM (Centre National de Recherches Meteorologiques).
}
\details{
Parametres :

NORTAV : Temperature moyenne journaliere (°C);

NORTNAV : Temperature minimale journaliere (°C);

NORTXAV : Temperature maximale journaliere (°C);

NORTRAV : Amplitude thermique journaliere (°C);

NORTXQ90 : Extreme chaud de la temperature maximale journaliere (°C);

NORTXQ10 : Extreme froid de la temperature maximale journaliere (°C);

NORTNQ10 : Extreme froid de la temperature minimale journaliere (°C);

NORTNQ90 : Extreme chaud de la temperature minimale journaliere (°C);

NORTXND : Nombre de jours anormalement chauds (jour);

NORTNND : Nombre de jours anormalement froids (jour);

NORTNHT : Nombre de nuits anormalement chaudes (jour);

NORTXHWD : Nombre de jours d_une vague de chaleur (jour);

NORTNCWD : Nombre de jours d_une vague de froid (jour);

NORTNFD : Nombre de jours de gel (jour);

NORTXFD : Nombre de jours sans degel (jour);

NORSD : Nombre de journees d_ete (jour);

NORTR : Nombre de nuits tropicales (jour);

NORHDD : Degre-jours de chauffage (°C);

NORCDD : Degre-jours de climatisation (°C);
}
\examples{
data(Temperatures)
}
\keyword{datasets}
